Added menu btn which hide actions. Changed search
This Burger button on hover show user actions menu with "log out" btn and "search" field;
After this menu appears, if you leave your cursor away from the header
the hide menu timer will be started. If you do not put the cursor again
on the header in 5 seconds , and if the "search" model is empty or === "", the actions
menu will become hidden and the burger menu will appear again.

Also, after searching, if the "search" input field been reseted to "",
or been closed , the hide menu timer will be started.